+++ src/main/java/in/techcamp/firstapp/KaikeiFinbookForm.java
@@ -7,7 +7,7 @@ import java.time.LocalDate;
 @Data
 public class KaikeiFinbookForm {
     private String name;
-    private long fiscalYear;
+    private int fiscalYear;
     private LocalDate startDate;
     private LocalDate endDate;
     private long groupId;
